+++ docs/Changelog.md
@@ -8,6 +8,7 @@
 * [#5438](https://github.com/TouK/nussknacker/pull/5438) [#5495](https://github.com/TouK/nussknacker/pull/5495) Improvement in DeploymentManager API:
     * Alignment in the api of primary (deploy/cancel) actions and the experimental api of custom actions.
 * [#5783](https://github.com/TouK/nussknacker/pull/5783) Added allowedProcessingModes field to components endpoint
+* [#5780](https://github.com/TouK/nussknacker/pull/5780) Fixed Scala case classes serialization when a class has additional fields in its body
 
 1.14.0 (21 Mar 2024)
 -------------------------